[S07P12B102-214] Feat. 회원관리: 토큰 재발급 코드 리팩토링 및 회원 인증 구현

+++ backend/src/main/java/com/jourgeois/backend/service/MemberService.java
@@ -75,7 +75,9 @@ public class MemberService {
                                     member.setNickname(data.getNickname());
                                     try{
                                         if(data.getProfileLink()!=null && !data.getProfileLink().isEmpty()){
-                                            s3Util.deleteFile(member.getProfileImg());
+                                            if(!member.getProfileImg().equals("default/1.png")) {
+                                                s3Util.deleteFile(member.getProfileImg());
+                                            }
                                             member.setProfileImg(s3Util.upload(data.getProfileLink(), data.getNickname()));
                                         }
                                     } catch (IOException e){
